+++ src/Adjustments/Adjusters/PercentDiscount.php
@@ -1,0 +1,69 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Adjustments\Adjusters;
+
+use Vanilo\Adjustments\Contracts\Adjustable;
+use Vanilo\Adjustments\Contracts\Adjuster;
+use Vanilo\Adjustments\Contracts\Adjustment;
+use Vanilo\Adjustments\Models\AdjustmentProxy;
+use Vanilo\Adjustments\Models\AdjustmentTypeProxy;
+use Vanilo\Adjustments\Support\HasWriteableTitleAndDescription;
+use Vanilo\Adjustments\Support\IsLockable;
+use Vanilo\Adjustments\Support\IsNotIncluded;
+
+final class PercentDiscount implements Adjuster
+{
+    use HasWriteableTitleAndDescription;
+    use IsLockable;
+    use IsNotIncluded;
+
+    private ?float $threshold;
+
+    public function __construct(
+        private float|int $percent
+    ) {
+    }
+
+    public static function reproduceFromAdjustment(Adjustment $adjustment): Adjuster
+    {
+        $data = $adjustment->getData();
+
+        return new self(floatval($data['percent'] ?? 0));
+    }
+
+    public function createAdjustment(Adjustable $adjustable): Adjustment
+    {
+        $adjustmentClass = AdjustmentProxy::modelClass();
+
+        return new $adjustmentClass($this->getModelAttributes($adjustable));
+    }
+
+    public function recalculate(Adjustment $adjustment, Adjustable $adjustable): Adjustment
+    {
+        $adjustment->setAmount($this->calculateAmount($adjustable));
+
+        return $adjustment;
+    }
+
+    private function calculateAmount(Adjustable $adjustable): float
+    {
+        return -1 * round($adjustable->preAdjustmentTotal() / 100  * $this->percent, 2);
+    }
+
+    private function getModelAttributes(Adjustable $adjustable): array
+    {
+        return [
+            'type' => AdjustmentTypeProxy::PROMOTION(),
+            'adjuster' => self::class,
+            'origin' => null,
+            'title' => $this->getTitle(),
+            'description' => $this->getDescription(),
+            'data' => ['percent' => $this->percent,],
+            'amount' => $this->calculateAmount($adjustable),
+            'is_locked' => $this->isLocked(),
+            'is_included' => $this->isIncluded(),
+        ];
+    }
+}