+++ Changelog.md
@@ -30,6 +30,7 @@
     - `CouponUtilized` (after a successful checkout converted to an order)
 - Added a listener to update coupon and promo usage after a `CouponUtilized` event (Foundation)
 - Added a promotion calculation listener (on CouponAdded and CouponRemoved events)
+- Added the `PercentDiscount` adjuster class
 
 ## 4.1.0
 ##### 2024-07-11